+++ src/Viserio/Component/Routing/Route.php
@@ -12,6 +12,7 @@ use Viserio\Component\Contracts\Container\Traits\ContainerAwareTrait;
 use Viserio\Component\Contracts\Routing\Route as RouteContract;
 use Viserio\Component\Routing\Traits\MiddlewareAwareTrait;
 use Viserio\Component\Support\Traits\InvokerAwareTrait;
+use Viserio\Component\Routing\Route\Parser as RouteParser;
 
 class Route implements RouteContract
 {
