+++ src/Viserio/Component/Routing/Tests/Generator/RouteTreeOptimizerTest.php
@@ -13,7 +13,7 @@ use Viserio\Component\Routing\Matchers\StaticMatcher;
 use Viserio\Component\Routing\TreeGenerator\ChildrenNodeCollection;
 use Viserio\Component\Routing\TreeGenerator\MatchedRouteDataMap;
 use Viserio\Component\Routing\TreeGenerator\RouteTreeNode;
-use Viserio\Component\Routing\TreeGenerator\RouteTreeOptimizer;
+use Viserio\Component\Routing\TreeGenerator\Optimizer\RouteTreeOptimizer;
 
 class RouteTreeOptimizerTest extends MockeryTestCase
 {
